requirements.txt
celery==5.3.1
Flask==2.2.2
flask_sqlalchemy==3.0.5
jsons==1.6.3
openai==0.26.4
pydantic==2.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.1
psycopg2-binary